Model changed rating to int32, unit of work implemented

+++ card_index_DAL/Entities/RateDetail.cs
@@ -10,6 +10,6 @@ namespace card_index_DAL.Entities
         public User User { get; set; }
         public int TextCardId { get; set; }
         public TextCard TextCard { get; set; }
-        public double RateValue { get; set; }
+        public ushort RateValue { get; set; }
     }
 }
